requirements.txt
fastapi
motor
dnspython
python-dotenv
google-generativeai
pydantic
crawl4ai
uvicorn